requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
ta==0.11.0
tensorflow==2.15.0
requests==2.31.0
python-dotenv==1.0.0
